pyproject.toml
[project]
name = "eebook-users"
version = "0.3.0"
description = "users microservice for the eEbook investment application, handling registration, authentication, and password management."
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "async-hvac>=0.7.3",
    "asyncpg>=0.30.0",
    "basedpyright>=1.36.1",
    "black>=25.12.0",
    "fastapi-jwt[authlib]>=0.3.0",
    "fastapi[standard]>=0.117.1",
    "hvac>=2.3.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "redis[hiredis]>=5.0.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "auto-changelog>=0.6.0",
    "bandit>=1.8.6",
    "detect-secrets>=1.5.0",
    "freezegun>=1.5.5",
    "greenlet>=3.2.4",
    "mkdocs>=1.6.1",
    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
    "mkdocs-material>=9.7.6",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "mkdocstrings[python]>=0.30.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.1",
    "types-hvac>=2.3.0.20250914",
    "types-passlib>=1.7.7.20250602",
    "types-setuptools>=80.9.0.20250822",
]

[tool.ruff]
exclude = [
    "*/migrations/*",
    "venv/",
    "__init__.py",
    "apps.py",
    "tests/*",
    "docs/**",
]
line-length = 100
indent-width = 4

[tool.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "W", "C", "F", "N", "COM", "I", "U", "N", "A", "M", "B", "D"]
ignore = ["F405", "W293", "D100", "D104", "D103", "D102", "D101", "D106", "D107", "D105", "B008", "E501"]
fixable = ["ALL"]
unfixable = []

[tool.bandit]
exclude = ["tests/"]

[tool.pytest]
asyncio_mode = "strict"
